Upgrading raml-validator-loader to v0.1.4
