+++ webpack/modules/raml-validator-loader/package.json
@@ -1,10 +1,10 @@
 {
   "name": "raml-validator-loader",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "Converts RAML document to javascript validation functions",
   "main": "lib/raml-validator-loader.js",
   "scripts": {
-    "test": "./node_modules/.bin/jest",
+    "test": "./node_modules/.bin/jest -c jest.config.json",
     "build": "./node_modules/.bin/webpack"
   },
   "author": "",
