+++ webpack/modules/raml-validator-loader/src/__tests__/RAMLValidator-test.js
@@ -1,34 +1,9 @@
-const RAML = require('raml-1-parser');
-const Generator = require('../Generator');
-const GeneratorContext = require('../GeneratorContext');
-
-/**
- * Utility function that parses RAML on-the-fly, generates a validator object
- * and returns the validation function
- */
-function createValidator(ramlDocument, options={}, typeName='TestType') {
-  var raml = RAML.parseRAMLSync(ramlDocument);
-  var types = raml.types().reduce(function(types, type) {
-    types[type.name()] = type;
-    return types;
-  }, {});
-
-  // Generate code with the given type
-  var ctx = new GeneratorContext(options);
-  ctx.uses( types[typeName].runtimeType() );
-
-  // Generate code
-  var code = Generator.generate(ctx);
-  var typeValidators = eval(code.replace('module.exports = ', ''));
-
-  // Return the validator for this type
-  typeValidators[typeName].code = code;
-  return typeValidators[typeName];
-}
+const ValidatorUtil = require('./utils/ValidatorUtil');
+const createValidator = ValidatorUtil.createValidator;
 
 describe('RAMLValidator', function () {
 
-  describe('Scalar Types', function () {
+  describe('Native Types', function () {
 
     describe('#nil', function () {
 
@@ -268,6 +243,11 @@ describe('RAMLValidator', function () {
       it('should validate if object', function () {
         var errors = this.validator({})
         expect(errors.length).toEqual(0);
+      });
+
+      it('should return error if null', function () {
+        var errors = this.validator(null)
+        expect(errors.length).toEqual(1);
       });
 
       it('should return error if number', function () {
